+++ Riot/Modules/KeyBackup/Setup/Intro/KeyBackupSetupIntroViewController.swift
@@ -133,12 +133,12 @@ final class KeyBackupSetupIntroViewController: UIViewController {
     private func showSkipAlert() {
         let alertController = UIAlertController(title: VectorL10n.keyBackupSetupSkipAlertTitle,
                                                 message: VectorL10n.keyBackupSetupSkipAlertMessage,
-                                                preferredStyle:.alert)
+                                                preferredStyle: .alert)
         
         alertController.addAction(UIAlertAction(title: VectorL10n.continue, style: .cancel, handler: { action in
         }))
         
-        alertController.addAction(UIAlertAction(title: VectorL10n.keyBackupSetupSkipAlertSkipAction , style: .default, handler: { action in
+        alertController.addAction(UIAlertAction(title: VectorL10n.keyBackupSetupSkipAlertSkipAction, style: .default, handler: { action in
             self.delegate?.keyBackupSetupIntroViewControllerDidCancel(self)
         }))